+++ netdef/Engines/webadmin/MyBaseView.py
@@ -1,4 +1,4 @@
-from flask import Flask, url_for, redirect, request
+from flask import Flask, url_for, redirect, request, abort
 
 from flask_admin import helpers, expose, model, BaseView
 import flask_login
@@ -7,4 +7,4 @@ class MyBaseView(BaseView):
     def is_accessible(self):
         return flask_login.current_user.is_authenticated
     def inaccessible_callback(self, name, **kwargs):
-        return redirect(url_for('/', next=request.url))
+        return abort(403)