+++ CHANGELOG.rst
@@ -4,6 +4,8 @@
 **Bug fixes**
 
 - Webadmin: Changed height of file edit textarea to 20 rows
+- Webadmin: Fixed routing.BuildError when you don't have the permission
+  to access the requested resource.
 
 
 1.0.4